Relocated config handling into its own subpackage
Allows config parsing and representation to be broken into separate modules.
