+++ salve/__init__.py
@@ -1,6 +1,6 @@
 from .class_resources import Enum, Singleton, with_metaclass
 from .log import Logger
 
-__version__ = '2.3.2'
+__version__ = '2.5.0'
 
 logger = Logger()
